+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Updated dependencies:
+    - cdc-util-0.21.0
+- `Config` data is now retrieved from Manifest.
+
+
 ## [0.21.0] - 2022-02-13
 ### Added
 - Added `CsvWriter` constructors using `OutputStream`.  #15